Added to Arrows and Scroll Bar
Got rid of the two arrows and started rotating the button instead of switching images, Gave a color transition to arrows when pointer hovers over them, also gave a slower transition to the center buttons for a smoother color change. Used webkit features to give the website a custom scroll bar with a radial gradient background similar to h1 and more coherent with the overall color scheme. Still need to debug the up and down arrows to make them understandable as the buttons refuse to rotate (webkit does not like changing aspects after defined which also made animation of the scroll bar impossible).